requirements.txt
pymongo>=3.11.3
discord.py>=1.6.0
fuzzywuzzy[speedup]>=0.18.0
